requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
psycopg2-binary==2.9.7
Pillow==10.0.0
python-decouple==3.8
